improve tests for CourseResource and ExerciseResource
also cleaned up some wrong implementations
allow to disable websocket logging during testing to solve weired test execution problems

+++ src/main/java/de/tum/in/www1/artemis/service/ProgrammingExerciseService.java
@@ -797,8 +797,12 @@ public class ProgrammingExerciseService {
 
         SolutionProgrammingExerciseParticipation solutionProgrammingExerciseParticipation = programmingExercise.getSolutionParticipation();
         TemplateProgrammingExerciseParticipation templateProgrammingExerciseParticipation = programmingExercise.getTemplateParticipation();
-        participationService.deleteResultsAndSubmissionsOfParticipation(solutionProgrammingExerciseParticipation.getId());
-        participationService.deleteResultsAndSubmissionsOfParticipation(templateProgrammingExerciseParticipation.getId());
+        if (solutionProgrammingExerciseParticipation != null) {
+            participationService.deleteResultsAndSubmissionsOfParticipation(solutionProgrammingExerciseParticipation.getId());
+        }
+        if (templateProgrammingExerciseParticipation != null) {
+            participationService.deleteResultsAndSubmissionsOfParticipation(templateProgrammingExerciseParticipation.getId());
+        }
         // This will also delete the template & solution participation.
         programmingExerciseRepository.delete(programmingExercise);
     }
